+++ twitch/api/communities.py
@@ -1,5 +1,6 @@
 from twitch.api.base import TwitchAPI
-from twitch.resources import Community
+from twitch.decorators import oauth_required
+from twitch.resources import Community, User
 
 
 class Communities(TwitchAPI):
@@ -33,3 +34,99 @@ class Communities(TwitchAPI):
             'email': email
         }
         self._request_put('communities/%s' % community_id, data=data)
+
+    def get_top(self, limit=10, cursor=None):
+        assert limit <= 100, 'Maximum number of videos returned in one request is 100'
+        params = {
+            'limit': limit,
+            'cursor': cursor
+        }
+        response = self._request_get('communities/top', params=params)
+        return [Community.construct_from(x) for x in response['communities']]
+
+    @oauth_required
+    def get_banned_users(self, community_id, limit=10, cursor=None):
+        assert limit <= 100, 'Maximum number of videos returned in one request is 100'
+
+        params = {
+            'limit': limit,
+            'cursor': cursor
+        }
+        response = self._request_get('communities/%s' % community_id, params=params)
+        return [User.construct_from(x) for x in response['banned_users']]
+
+    @oauth_required
+    def ban_user(self, community_id, user_id):
+        self._request_put('communities/%s/bans/%s' % (community_id, user_id))
+
+    @oauth_required
+    def unban_user(self, community_id, user_id):
+        self._request_delete('communities/%s/bans/%s' % (community_id, user_id))
+
+    @oauth_required
+    def create_avatar_image(self, community_id, avatar_image):
+        data = {
+            'avatar_image': avatar_image,
+        }
+        self._request_post('communities/%s/images/avatar' % community_id, data=data)
+
+    @oauth_required
+    def delete_avatar_image(self, community_id, avatar_image):
+        self._request_delete('communities/%s/images/avatar' % community_id)
+
+    @oauth_required
+    def create_cover_image(self, community_id, cover_image):
+        data = {
+            'cover_image': cover_image,
+        }
+        self._request_post('communities/%s/images/cover' % community_id, data=data)
+
+    @oauth_required
+    def delete_cover_image(self, community_id):
+        self._request_delete('communities/%s/images/cover' % community_id)
+
+    def get_moderators(self, community_id):
+        response = self._request_get('communities/%s/moderators' % community_id)
+        return [User.construct_from(x) for x in response['moderators']]
+
+    @oauth_required
+    def add_moderator(self, community_id, user_id):
+        self._request_put('communities/%s/moderators/%s' % (community_id, user_id))
+
+    @oauth_required
+    def delete_moderator(self, community_id, user_id):
+        self._request_delete('communities/%s/moderators/%s' % (community_id, user_id))
+
+    @oauth_required
+    def get_permissions(self, community_id):
+        response = self._request_get('communities/%s/permissions' % community_id)
+        return response
+
+    @oauth_required
+    def report_violation(self, community_id, channel_id):
+        data = {
+            'channel_id': channel_id,
+        }
+        self._request_post('communities/%s/report_channel' % community_id, data=data)
+
+    @oauth_required
+    def get_timed_out_users(self, community_id, limit=10, cursor=None):
+        assert limit <= 100, 'Maximum number of videos returned in one request is 100'
+        params = {
+            'limit': limit,
+            'cursor': cursor
+        }
+        response = self._request_get('communities/%s/timeouts' % community_id, params=params)
+        return [User.construct_from(x) for x in response['timed_out_users']]
+
+    @oauth_required
+    def add_timed_out_user(self, community_id, user_id, duration, reason=None):
+        data = {
+            'duration': duration,
+            'reason': reason,
+        }
+        self._request_put('communities/%s/timeouts/%s' % (community_id, user_id), data=data)
+
+    @oauth_required
+    def delete_timed_out_user(self, community_id, user_id):
+        self._request_delete('communities/%s/timeouts/%s' % (community_id, user_id))